requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
django-phonenumber-field[phonenumberslite]
Babel
google-auth
Pillow